exclude ancestors from summary id
