+++ internal/prefix/constant.go
@@ -1,0 +1,5 @@
+package prefix
+
+const SvrAncestor = "svr.ancestor"
+const SvrChild = "svr.child"
+const SvrParent = "svr.parent"
